# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medical Records API",
        "description": "API quan ly ho so y te",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/medical-records": {
            "get": {
                "tags": [
                    "Medical Records"
                ],
                "summary": "Danh sách hồ sơ bệnh án",
                "operationId": "0de61bef7f027788980e40206c5daf5f",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "doctor_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "appointment_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "primary_icd",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "scheduled",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách hồ sơ bệnh án"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "post": {
                "tags": [
                    "Medical Records"
                ],
                "summary": "Tạo hồ sơ bệnh án mới",
                "operationId": "c8356fc1fbf04bbdd7aba0e62f29938e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_id": {
                                        "type": "string",
                                        "example": "patient_123"
                                    },
                                    "doctor_id": {
                                        "type": "string",
                                        "example": "doctor_456"
                                    },
                                    "visit_info.visit_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "visit_info.visit_type": {
                                        "type": "string",
                                        "example": "consultation"
                                    },
                                    "visit_info.chief_complaint": {
                                        "type": "string",
                                        "example": "Đau đầu"
                                    },
                                    "visit_info.appointment_id": {
                                        "type": "string",
                                        "example": "appointment_789"
                                    },
                                    "examination": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "diagnosis": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "treatment_plan": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "scheduled",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/v1/medical-records/{id}": {
            "get": {
                "tags": [
                    "Medical Records"
                ],
                "summary": "Chi tiết hồ sơ bệnh án",
                "operationId": "22ebbabaa794e910a6a8c2835f329ee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin hồ sơ bệnh án"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "put": {
                "tags": [
                    "Medical Records"
                ],
                "summary": "Cập nhật hồ sơ bệnh án",
                "operationId": "7d8f1196f50a3cd06583a783d60c68d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_rev": {
                                        "type": "string",
                                        "example": "1-abc123"
                                    },
                                    "examination": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "diagnosis": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "scheduled",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Medical Records"
                ],
                "summary": "Xóa hồ sơ bệnh án",
                "operationId": "ff52fff3926a57d80cb68ddf2ac36ad5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rev",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "400": {
                        "description": "Thiếu rev parameter"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
...
    },
...
}